refactor: move dgraph independent code to utils
