+++ pkg/controller/dgraph/models/pv.go
@@ -19,7 +19,6 @@ package models
 
 import (
 	"github.com/Sirupsen/logrus"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"time"
 
@@ -33,7 +32,6 @@ import (
 // Dgraph Model Constants
 const (
 	IsPersistentVolume = "isPersistentVolume"
-	StorageDefault     = "purser-default"
 )
 
 // PersistentVolume schema in dgraph
@@ -58,7 +56,7 @@ func createPersistentVolumeObject(pv api_v1.PersistentVolume, client *kubernetes
 	}
 	capacity := pv.Spec.Capacity["storage"]
 	newPv.StorageCapacity = utils.ConvertToFloat64GB(&capacity)
-	newPv.StorageType = getStorageType(pv, client)
+	newPv.StorageType = utils.GetFinalStorageTypeOfPV(pv, client)
 	logrus.Debugf("PV: %s, storageType: %s", newPv.Name, newPv.StorageType)
 
 	deletionTimestamp := pv.GetDeletionTimestamp()
@@ -107,44 +105,4 @@ func CreateOrGetPersistentVolumeByID(xid string) string {
 		return ""
 	}
 	return assigned.Uids["blank-0"]
-}
-
-/* getStorageType
-   input: persistent volume
-   output: the type(final) of PV's storage class
-   i.e., if PV has storage class A, A is of type B(storage class) and so on..
-   until a storage class X is of its own type X. Then this function returns the final type of PV's storage as X
-
-   "purser-default" is returned in special cases:
-   1. if A is of type B, if B is of type A (i.e., if a cycle is found)
-   2. an error is encountered
-   3. if A is not having any type i.e., "" (empty string case)
-*/
-func getStorageType(pv api_v1.PersistentVolume, client *kubernetes.Clientset) string {
-	cycleChecker := make(map[string]bool)
-	logrus.Debugf("PV: %s, storageClass: %s", pv.Name, pv.Spec.StorageClassName)
-	return getFinalTypeOfStorageClass(client, pv.Spec.StorageClassName, cycleChecker)
-}
-
-// getFinalTypeOfStorageClass
-// this is helper function for func getStorageType
-func getFinalTypeOfStorageClass(client *kubernetes.Clientset, storageClassName string, cycleChecker map[string]bool) string {
-	if _, isVisited := cycleChecker[storageClassName]; isVisited {
-		return StorageDefault
-	} else {
-		cycleChecker[storageClassName] = true
-	}
-
-	storageClass, err := utils.RetrieveStorageClass(client, metav1.GetOptions{}, storageClassName)
-	if err != nil {
-		return StorageDefault
-	}
-
-	storageType := storageClass.Parameters["type"]
-	if storageType == "" {
-		return StorageDefault
-	} else if storageType == storageClassName {
-		return storageClassName
-	}
-	return getFinalTypeOfStorageClass(client, storageType, cycleChecker)
 }
